Datathon_Fase_5/requirements.txt
pandas
numpy
scikit-learn
imblearn
xgboost
spacy
azure-storage-blob
tqdm
langdetect
deep-translator
streamlit

pt_core_news_md @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_md-3.7.0/pt_core_news_md-3.7.0-py3-none-any.whl
